nixing default singleton for non-fatal exits
